+++ build.gradle
@@ -40,7 +40,7 @@ def jsonProviders = [
          group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.18.0'],
         [id   : 'gson', additionalVersions: ['2.3'],
          group: 'com.google.code.gson', name: 'gson', version: '2.11.0'],
-        [id   : 'orgJson', additionalVersions: [],
+        [id   : 'orgJson', additionalVersions: ['20201115'],
          group: 'org.json', name: 'json', version: '20240303'],
         [id   : 'kotlinxJson', additionalVersions: [],
          group: 'org.jetbrains.kotlinx', name: 'kotlinx-serialization-json', version: '1.7.3'],
