--- src/integration/orgJsonTest/IsolatedOrgJsonTest.java
@@ -1,25 +1,0 @@
-import dev.harrel.jsonschema.ValidatorFactory;
-import dev.harrel.jsonschema.providers.OrgJsonNode;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-
-class IsolatedOrgJsonTest {
-    @Test
-    void shouldInstantiateValidatorFactory() {
-        new ValidatorFactory();
-    }
-
-    @Test
-    void shouldPassForOrgJsonFactory() {
-        new ValidatorFactory()
-                .withJsonNodeFactory(new OrgJsonNode.Factory())
-                .validate("{}", "{}");
-    }
-
-    @Test
-    void shouldFailForDefaultFactory() {
-        assertThatThrownBy(() -> new ValidatorFactory().validate("{}", "{}"))
-                .isInstanceOf(NoClassDefFoundError.class);
-    }
-}
